[refactor] errormessage 패키징 리팩토링(#10)
