+++ src/main/java/org/recordy/server/auth/exception/UnauthorizedException.java
@@ -1,5 +1,6 @@
 package org.recordy.server.auth.exception;
 
+import org.recordy.server.auth.message.ErrorMessage;
 import org.recordy.server.common.exception.RecordyException;
 
 public class UnauthorizedException extends RecordyException {
