+++ src/main/java/org/recordy/server/auth/service/impl/kakao/AuthKakaoPlatformServiceImpl.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.recordy.server.auth.domain.AuthPlatform;
 import org.recordy.server.auth.domain.usecase.AuthSignIn;
-import org.recordy.server.auth.exception.ErrorMessage;
+import org.recordy.server.auth.message.ErrorMessage;
 import org.recordy.server.auth.exception.UnauthorizedException;
 import org.recordy.server.auth.kakao.KakaoOAuthProvider;
 import org.recordy.server.auth.service.AuthPlatformService;
