+++ src/main/java/org/recordy/server/common/message/ErrorMessage.java
@@ -1,0 +1,4 @@
+package org.recordy.server.common.message;
+
+public enum ErrorMessage {
+}
